Split validate_types file into multiple classes
Creates separate files for the ValidationState, Function and
BasicBlock classes.
