+++ source/validate.cpp
@@ -24,6 +24,8 @@
 // TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 // MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
 
+#include <validate.h>
+
 #include <cassert>
 #include <cstdio>
 
@@ -34,17 +36,17 @@
 #include <string>
 #include <vector>
 
-#include "validate.h"
-#include "validate_passes.h"
-
-#include "binary.h"
-#include "diagnostic.h"
-#include "instruction.h"
-#include "opcode.h"
-#include "operand.h"
-#include "spirv-tools/libspirv.h"
-#include "spirv_constant.h"
-#include "spirv_endian.h"
+#include <binary.h>
+#include <diagnostic.h>
+#include <instruction.h>
+#include <opcode.h>
+#include <operand.h>
+#include <spirv-tools/libspirv.h>
+#include <spirv_constant.h>
+#include <spirv_endian.h>
+#include <val/Construct.h>
+#include <val/Function.h>
+#include <val/ValidationState.h>
 
 using std::function;
 using std::ostream_iterator;
@@ -123,6 +125,7 @@ void DebugInstructionPass(ValidationState_t& _,
     default:
       break;
   }
+
 }
 
 // Collects use-def info about an instruction's IDs.
