+++ source/validate_cfg.cpp
@@ -24,8 +24,11 @@
 // TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 // MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
 
-#include "validate.h"
-#include "validate_passes.h"
+#include <validate.h>
+#include <val/BasicBlock.h>
+#include <val/Construct.h>
+#include <val/Function.h>
+#include <val/ValidationState.h>
 
 #include <algorithm>
 #include <cassert>
@@ -242,7 +245,7 @@ spv_result_t PerformCfgChecks(ValidationState_t& _) {
     }
 
     // Check all headers dominate their merge blocks
-    for (CFConstruct& construct : function.get_constructs()) {
+    for (Construct& construct : function.get_constructs()) {
       auto header = construct.get_header();
       auto merge = construct.get_merge();
       // auto cont = construct.get_continue();
