+++ test/ValidationState.cpp
@@ -30,9 +30,12 @@
 #include <gtest/gtest.h>
 #include <vector>
 
-#include "spirv/spirv.h"
+#include <spirv/spirv.h>
+#include <val/Construct.h>
+#include <val/Function.h>
+#include <val/ValidationState.h>
 
-#include "source/validate.h"
+#include <validate.h>
 
 namespace {
 using libspirv::ValidationState_t;
